app/requirements.txt
Flask 
ufal.udpipe 
conllu 
